Refactor: service interface 삭제
